+++ front/src/components/pages/BlogPost/BlogPostContainer.tsx
