+++ front/src/components/pages/BlogPost/BlogPostPresenter.tsx
