+++ front/src/components/pages/BlogPost/index.ts
